updat

+++ database/migrations/2024_05_27_161109_create_order_details_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->integer('order_id');
             $table->integer('customer_id');
             $table->integer('product_id');
-            $table->integer('product_price');
+            $table->float('product_price');
             $table->integer('product_qty');
             $table->timestamps();
         });
